hilda_garde/requirements/common.txt
discord==1.0.1
python-decouple==3.1
flask==1.1.2
